feature: Multiple titles
Hi there,

Thanks very much for your work on mulitplot.  I found it really useful.  I added a feature so that it was possible to add multiple titles.  I wanted to be able to label panels for a complex figure.  Feel free to use/discard/adjust.

I could not neatly find a way to left justify within a panel (i used hjust in another version, but it needs to be set for the dims of the specific plot).  If you figure it out, i think that would be quite handy.

Den